Hack in (special) character support
I deliberately choose "hack" instead of "add" here, since there is no
unicode support, while at the same time we assume that Dafny runs with
the default (?) `--unicode-char:true`.

Additionally, we probably do not handle all ASCII escape characters,
and I don't really know what kind of encoding in each step of the
pipeline from Dafny to PolyML/HOL is used (or whether that actually matters).

Finally, I am unsure on when/what to escape. The goal for now is to
implement this in a way such that `print` statements match what `dafny
run` would do.